added `findMedia()` query
